+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * fix selection menu in latest Android
 * fix wrong number of chars per line when insert called after init and scrollbar appear
 * fix missing cursor when init cmd plugin (without calling refresh)
+* fix issue with function prompt not updating after resume
 
 ## 1.23.2
 ### Bugfix